+++ src/js/render/index.mjs
@@ -1,0 +1,4 @@
+export * from "./followers.mjs";
+export * from "./suggestions.mjs";
+export * from "./profile.mjs";
+export * from "./feed.mjs";
